Refactor(foldering): Enum 파일들을 domain package가 아닌 constant package로 이동
constant package 생성 및 파일 이동
