+++ src/main/java/christmas/domain/OrderItem.java
@@ -1,5 +1,7 @@
 package christmas.domain;
 
+import christmas.constants.Menu;
+
 import java.util.Arrays;
 
 public class OrderItem {
