+++ src/main/java/christmas/domain/Orders.java
@@ -1,5 +1,7 @@
 package christmas.domain;
 
+import christmas.constants.Menu;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
