+++ src/main/java/christmas/domain/PromotionManager.java
@@ -1,5 +1,8 @@
 package christmas.domain;
 
+import christmas.constants.Badge;
+import christmas.constants.Menu;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
